JVM IR: remove MetadataSource.Class, use attributes of IrAttributeContainer instead
Store the original ClassDescriptor (before lowerings) for IrClass in a
hash map and pass it to ClassCodegen.

Both mechanisms (MetadataSource and IrAttributeContainer-backed maps)
essentially do the same thing, but the latter is more tested because of
the usages related to LocalClassInfo. So we're going to deprecate and
remove MetadataSource eventually, replacing it with
IrAttributeContainer-based solution.

The difference in behavior here was that metadataSource was not copied
in DeepCopy, so metadata of local classes was lost whenever the class
was "moved" to a new place by invoking `deepCopyWithSymbols`. In
particular, deep copy was used in InitializersLowering. In the fixed
tests, the copied anonymous objects did not have any metadata and were
generated as synthetic classes, which in turn made frontend crash on
loading such classes.

Some notes:
1) The fact that frontend loads local declarations from other modules is
   scary and probably should be fixed in some way in the future. We do
   need to have local types in the Kotlin metadata for reflection use
   cases, but the compiler should approximate them to avoid looking for
   classes with potentially unstable names.
2) InitializersLowering in fact does a deep copy of every field
   initializer into every constructor of the class. So the same local
   class may be present multiple times in the IR, which is incorrect.
   However, on JVM we generate classes with fixed names (computed by
   InventNamesForLocalClasses), and since we _do_ copy attributes in
   deep copy, all the copied instances will get the same name and
   collapse into a single .class file. Now, they will also get the same
   metadata too. This is basically magic and we shouldn't rely on this.
   InitializersLowering needs to "pull up" local classes from field
   initializers to declare each of them only once, however doing that is
   not trivial, so it's postponed.