(TODO) JVM: remove -Xno-unified-null-checks
TODO: ensure there are tests on bytecode transformations that fail if we
accidentally remove support for old Intrinsics methods.

This was a fallback flag to disable behavior that was implemented in
Kotlin 1.4.

We still need compiler tests with code compiled without unified null
checks though, so we'll use the "// API_VERSION: 1.3" directive instead.