Reflection: add a system property to control reflect implementation
Also add tests to run the legacy implementation on
`codegen/box/reflection`. It's not the whole `codegen/box` directory
because it would be wasteful to run so many tests, which are not
affected by reflect, in a separate process. And a separate process is
needed to set the system property.

 #KT-80203 Fixed